Move PathPair listing functionality back to file_lister.py as function
