+++ xfel/merging/application/input/file_load_calculator.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 from six.moves import range
 import sys
 import os
-from xfel.merging.application.input.path_pair import PathPairList
+from xfel.merging.application.input.file_lister import list_input_pairs
 
 debug=False # set debug=True, if you want to see a per-rank experiments/reflections file pair list generated by this calculator
 available_rank_count = 0  # used only when running this script stand-alone:
@@ -124,7 +124,7 @@ class Script(Script_Base):
 
   def get_file_list(self):
     '''Get experiments/reflections file list'''
-    file_list = PathPairList(self.params)
+    file_list = list_input_pairs(self.params)
     print("Built an input list of %d experiments/reflections file pairs"%len(file_list))
     print("To view the list, set debug=True in file_load_calculator.py")
     return file_list
